Add files via upload
# Simon-says-game
This is a simon says game but instead of people giving words to each other 4 colours are there. This game helps one build their memory. On each level a pattern of colour is given and he/she needs to remember the pattern and click on each block to complete the pattern.